Refactor feature helpers into several modules
