+++ spec/support/feature_helpers/sessions.rb
@@ -1,22 +1,11 @@
 module FeatureHelpers
   module Sessions
     def login
-      email = 'person@prison.com'
-      password = 'secret123'
-      create_user(email, password)
+      create :user
       visit root_path
-      fill_in 'user[email]', with: email
-      fill_in 'user[password]', with: password
+      fill_in 'user[email]', with: attributes_for(:user)[:email]
+      fill_in 'user[password]', with: attributes_for(:user)[:password]
       click_button 'Log in'
-    end
-
-    def create_user(email, password)
-      create(
-        :user,
-        email: email,
-        password: password,
-        password_confirmation: password
-      )
     end
 
     def sign_out
@@ -33,6 +22,19 @@ module FeatureHelpers
       fill_in 'user[password]', with: password
       fill_in 'user[password_confirmation]', with: password
       click_button 'Set my password'
+    end
+
+    def start_password_reset
+      create :user
+      visit new_user_password_path
+      fill_in 'user[email]', with: attributes_for(:user)[:email]
+      click_button 'Reset password'
+    end
+
+    def change_password(new_password)
+      fill_in 'user[password]', with: new_password
+      fill_in 'user[password_confirmation]', with: new_password
+      click_button 'Change my password'
     end
   end
 end
